Responsive header
Responsive fixed header that consists of dynamic links that change when a user is logged in or not.

On smaller viewports links are replaced with an animated burger menu. This works with CSS and Javascript.

Javascript adds an overlay class, which will be made later on in the project. Defending on the content that will be needed.