Register, Authentication & Sign-out
Register: ability to register accounts by submitting a form.
Functionality is working correctly. Form will be added to the modal later on when styling is final (home and sign-in page).

First user created to use for further development.

Authenticate: Functionality to authenticate and start a session from form input on the signin.php page. Redirects the user to his/her profile page. Styling the profile page will be done later.

Logout: Functionality is working.

Small changes to paths and adding the bare medium of the modal.