Laying out the profile page.
Laying out the content on the profile page. Also completed the styling so database user results can easily be placed onto the page.

Small change has been made to the header. It now displays the logout button (was supposed to be on the profile page). Turns out this would be better for accessability.

Database results will be added later on.