function profile page & multi-step form.
To finalize the profile page certain user specific information has to be shown.

Function to retrieve this data based on the current logged in session is made and will be used to display this data later on.

Also implemented a form wizard to improve user experience when creating an account.

Multi-step form made with JavaScript that validates whether a input is empty or not. It also displays a progress indication.

This finalizes the implementation of registering accounts.