Handling database lesson results.
Constructing and implementing the 'getLessons' function.

This function displays database results (lessons) as a list item until the starting DATETIME is reached.

Buttons inside the result change based on the logged in user. Members have the ability to join a lesson, employees are able to check who joined. This functionality will be added later on and only needs small changes to the anchor tags.

Finished styling of the timeline and it's results.

Small changes made for responsiveness inside the header.