Implementing ability to register for a lesson, and retrieve a list of participants.
The rooster.php page gives members the possibility to register for a lesson by clicking on 'inschrijven'. Functionality behind the button has been implemented. On click their and the lessons ID will be placed in the 'gebruikers_has_lessen' table.

When logged in as an employee, the button 'deelnemerslijst' will be visible. Any employee will be able to fetch a list of members that signed up for that specific lesson.

The list will be displayed on 'participants.php'.